dataset-collection/requirements.txt
pymongo
requests
beautifulsoup4
Pillow